src/abyss-intelligence/pyproject.toml
[project]
name = "abyss-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx[http2]>=0.28.1",
    "mcp>=1.24.0",
    "python-dotenv>=1.2.1",
    "surrealdb>=1.0.7",
]
